DON-1205: Move location search test to separate file before adding more scenarios
