+++ integrationTests/CampaignRepositoryTest.php
@@ -190,91 +190,6 @@ class CampaignRepositoryTest extends IntegrationTest
         $this->assertSame(['Campaign Active', 'Campaign Expired', 'Campaign Preview'], $returnCampaignNames);
     }
 
-    public function testSearchSortsByLocation(): void
-    {
-        $sut = $this->getService(CampaignRepository::class);
-        $em = $this->getService(EntityManagerInterface::class);
-
-        $charity = TestCase::someCharity();
-
-        $campaign = $this->createCampaign(
-            charity: $charity,
-            name: 'Campaign For UK',
-            status: CampaignStatus::Active,
-            withUniqueSalesforceId: true,
-            metaCampaignSlug: 'some-slug',
-            relatedApplicationStatus: ApplicationStatus::Approved,
-            relatedApplicationCharityResponseToOffer: CharityResponseToOffer::Accepted
-        );
-
-        $campaign->replaceLocations([
-            ['countryName' => 'uk', 'regionCode' => null],
-        ]);
-        $stats = CampaignStatistics::zeroPlaceholder($campaign, new \DateTimeImmutable('now'));
-
-        $em->persist($campaign);
-        $em->persist($stats);
-
-        $haringeyCampaign = $this->createCampaign(
-            charity: $charity,
-            name: 'Campaign For Haringey',
-            status: CampaignStatus::Active,
-            withUniqueSalesforceId: true,
-            metaCampaignSlug: 'some-slug',
-            relatedApplicationStatus: ApplicationStatus::Approved,
-            relatedApplicationCharityResponseToOffer: CharityResponseToOffer::Accepted
-        );
-
-        $haringeyCampaign->replaceLocations([
-            ['countryName' => 'uk', 'regionCode' => null],
-            ['countryName' => null, 'regionCode' => 'E09000014'],
-        ]);
-        $stats = CampaignStatistics::zeroPlaceholder($haringeyCampaign, new \DateTimeImmutable('now'));
-
-        $em->persist($haringeyCampaign);
-        $em->persist($stats);
-
-        $londonCampaign = $this->createCampaign(
-            charity: $charity,
-            name: 'Campaign For London',
-            status: CampaignStatus::Active,
-            withUniqueSalesforceId: true,
-            metaCampaignSlug: 'some-slug',
-            relatedApplicationStatus: ApplicationStatus::Approved,
-            relatedApplicationCharityResponseToOffer: CharityResponseToOffer::Accepted
-        );
-
-        $londonCampaign->replaceLocations([
-            ['countryName' => 'uk', 'regionCode' => null],
-            ['countryName' => null, 'regionCode' => 'E12000007'],
-        ]);
-        $stats = CampaignStatistics::zeroPlaceholder($londonCampaign, new \DateTimeImmutable('now'));
-
-        $em->persist($londonCampaign);
-        $em->persist($stats);
-
-        $em->flush();
-
-        $returnValue = $sut->search(
-            sortField: 'location',
-            sortDirection: 'desc',
-            regions: ['E09000014','E12000007', 'E92000001'],
-            offset: 0,
-            limit: 6,
-            metaCampaignSlug: 'some-slug',
-            fundSlug: null,
-            jsonMatchInListConditions: [],
-            term: null,
-        );
-
-        $returnCampaignNames = array_map(
-            static fn(Campaign $campaign) => $campaign->getCampaignName(),
-            $returnValue
-        );
-
-        $this->assertSame(['Campaign For Haringey', 'Campaign For London', 'Campaign For UK'], $returnCampaignNames);
-    }
-
     private function getCharityAwaitingGiftAidApproval(): Charity
     {
         $charity = TestCase::someCharity();
